Move track criterium stuff to the common directory
Move trackcriteria.* and trackcriteriumevaluation.* from the
"desktop-remote" directory to the "common" directory.

This will make it easier to write unit tests. And in the future we
will probably also use these expression trees in the server anyway.
