+++ src/common/trackcriteria.cpp
