+++ src/common/trackcriteria.h
