+++ src/common/trackcriteriumevaluation.cpp
