+++ src/desktop-remote/desktop-remote-metatypes.cpp
@@ -1,5 +1,5 @@
 /*
-    Copyright (C) 2020-2025, Kevin André <hyperquantum@gmail.com>
+    Copyright (C) 2020-2026, Kevin André <hyperquantum@gmail.com>
 
     This file is part of PMP (Party Music Player).
 
@@ -17,8 +17,6 @@
     with PMP.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "trackcriteria.h"
-
 #include <QMetaType>
 
 namespace PMP
@@ -30,7 +28,7 @@ namespace PMP
     protected:
         DesktopRemoteMetatypesInit()
         {
-            qRegisterMetaType<PMP::PredefinedTrackCriterium>();
+            // no metatypes to register at this time
         }
 
     private:
